working program

+++ bamazon_DB.sql
@@ -7,7 +7,7 @@ CREATE TABLE products (
     item_id INT NOT NULL AUTO_INCREMENT,
     product_name VARCHAR(30) NOT NULL,
     department_name VARCHAR(30) NOT NULL,
-    price INT default 0,
+    price DECIMAL (10,2) default 0,
     stock_quantity INT default 0,
     PRIMARY KEY (item_id)
 );
